BugFix in plams-ams interface: inconsistent units between atomic and lattice vector coords
The read_molecule method gets both the atomic coordinates and the lattice vectors in bohr from the rkf file.  However, the atomic coordinates are converted to angstrom by the Atom constructor, requiring the lattice vectors to be converted as well.

Now the atomic and vector coordinates stored in the Molecule object returned by read_molecule will both be in angstrom.